+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/api/healthDataAIServicesContext.ts
@@ -3,23 +3,34 @@
 
 import { logger } from "../logger.js";
 import { KnownVersions } from "../models/models.js";
+import { AzureSupportedClouds, getArmEndpoint } from "../static-helpers/cloudSettingHelpers.js";
 import { Client, ClientOptions, getClient } from "@azure-rest/core-client";
 import { TokenCredential } from "@azure/core-auth";
 
-export interface HealthDataAIServicesContext extends Client {}
+export interface HealthDataAIServicesContext extends Client {
+  /** The ID of the target subscription. The value must be an UUID. */
+  subscriptionId: string;
+  /** The API version to use for this operation. */
+  /** Known values of {@link KnownVersions} that the service accepts. */
+  apiVersion?: string;
+}
 
 /** Optional parameters for the client. */
 export interface HealthDataAIServicesClientOptionalParams extends ClientOptions {
   /** The API version to use for this operation. */
   /** Known values of {@link KnownVersions} that the service accepts. */
   apiVersion?: string;
+  /** Specifies the Azure cloud environment for the client. */
+  cloudSetting?: AzureSupportedClouds;
 }
 
 export function createHealthDataAIServices(
   credential: TokenCredential,
+  subscriptionId: string,
   options: HealthDataAIServicesClientOptionalParams = {},
 ): HealthDataAIServicesContext {
-  const endpointUrl = options.endpoint ?? options.baseUrl ?? `https://management.azure.com`;
+  const endpointUrl =
+    options.endpoint ?? getArmEndpoint(options.cloudSetting) ?? "https://management.azure.com";
   const prefixFromOptions = options?.userAgentOptions?.userAgentPrefix;
   const userAgentInfo = `azsdk-js-arm-healthdataaiservices/1.0.0-beta.1`;
   const userAgentPrefix = prefixFromOptions
@@ -29,27 +40,9 @@ export function createHealthDataAIServices(
     ...options,
     userAgentOptions: { userAgentPrefix },
     loggingOptions: { logger: options.loggingOptions?.logger ?? logger.info },
-    credentials: {
-      scopes: options.credentials?.scopes ?? [`${endpointUrl}/.default`],
-    },
+    credentials: { scopes: options.credentials?.scopes ?? [`${endpointUrl}/.default`] },
   };
   const clientContext = getClient(endpointUrl, credential, updatedOptions);
-  clientContext.pipeline.removePolicy({ name: "ApiVersionPolicy" });
-  const apiVersion = options.apiVersion ?? "2024-09-20";
-  clientContext.pipeline.addPolicy({
-    name: "ClientApiVersionPolicy",
-    sendRequest: (req, next) => {
-      // Use the apiVersion defined in request url directly
-      // Append one if there is no apiVersion and we have one at client options
-      const url = new URL(req.url);
-      if (!url.searchParams.get("api-version")) {
-        req.url = `${req.url}${
-          Array.from(url.searchParams.keys()).length > 0 ? "&" : "?"
-        }api-version=${apiVersion}`;
-      }
-
-      return next(req);
-    },
-  });
-  return clientContext;
+  const apiVersion = options.apiVersion;
+  return { ...clientContext, apiVersion, subscriptionId } as HealthDataAIServicesContext;
 }