+++ sdk/healthdataaiservices/arm-healthdataaiservices/src/classic/deidServices/index.ts
@@ -3,42 +3,36 @@
 
 import { HealthDataAIServicesContext } from "../../api/healthDataAIServicesContext.js";
 import {
-  deidServicesGet,
-  deidServicesListByResourceGroup,
-  deidServicesListBySubscription,
-  deidServicesCreate,
-  deidServicesUpdate,
-  deidServicesDelete,
-} from "../../api/deidServices/index.js";
-import { DeidService, DeidUpdate } from "../../models/models.js";
+  $delete,
+  create,
+  listBySubscription,
+  listByResourceGroup,
+  get,
+} from "../../api/deidServices/operations.js";
+import {
+  DeidServicesDeleteOptionalParams,
+  DeidServicesCreateOptionalParams,
+  DeidServicesListBySubscriptionOptionalParams,
+  DeidServicesListByResourceGroupOptionalParams,
+  DeidServicesGetOptionalParams,
+} from "../../api/deidServices/options.js";
+import { DeidService } from "../../models/models.js";
 import { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
 import { PollerLike, OperationState } from "@azure/core-lro";
-import {
-  DeidServicesGetOptionalParams,
-  DeidServicesListByResourceGroupOptionalParams,
-  DeidServicesListBySubscriptionOptionalParams,
-  DeidServicesCreateOptionalParams,
-  DeidServicesUpdateOptionalParams,
-  DeidServicesDeleteOptionalParams,
-} from "../../api/options.js";
 
 /** Interface representing a DeidServices operations. */
 export interface DeidServicesOperations {
-  /** Get a DeidService */
-  get: (
+  /** Delete a DeidService */
+  /**
+   *  @fixme delete is a reserved word that cannot be used as an operation name.
+   *         Please add @clientName("clientName") or @clientName("<JS-Specific-Name>", "javascript")
+   *         to the operation to override the generated name.
+   */
+  delete: (
     resourceGroupName: string,
     deidServiceName: string,
-    options?: DeidServicesGetOptionalParams,
-  ) => Promise<DeidService>;
-  /** List DeidService resources by resource group */
-  listByResourceGroup: (
-    resourceGroupName: string,
-    options?: DeidServicesListByResourceGroupOptionalParams,
-  ) => PagedAsyncIterableIterator<DeidService>;
-  /** List DeidService resources by subscription ID */
-  listBySubscription: (
-    options?: DeidServicesListBySubscriptionOptionalParams,
-  ) => PagedAsyncIterableIterator<DeidService>;
+    options?: DeidServicesDeleteOptionalParams,
+  ) => PollerLike<OperationState<void>, void>;
   /** Create a DeidService */
   create: (
     resourceGroupName: string,
@@ -46,75 +40,54 @@ export interface DeidServicesOperations {
     resource: DeidService,
     options?: DeidServicesCreateOptionalParams,
   ) => PollerLike<OperationState<DeidService>, DeidService>;
-  /** Update a DeidService */
-  update: (
+  /** List DeidService resources by subscription ID */
+  listBySubscription: (
+    options?: DeidServicesListBySubscriptionOptionalParams,
+  ) => PagedAsyncIterableIterator<DeidService>;
+  /** List DeidService resources by resource group */
+  listByResourceGroup: (
+    resourceGroupName: string,
+    options?: DeidServicesListByResourceGroupOptionalParams,
+  ) => PagedAsyncIterableIterator<DeidService>;
+  /** Get a DeidService */
+  get: (
     resourceGroupName: string,
     deidServiceName: string,
-    properties: DeidUpdate,
-    options?: DeidServicesUpdateOptionalParams,
-  ) => PollerLike<OperationState<DeidService>, DeidService>;
-  /** Delete a DeidService */
-  delete: (
-    resourceGroupName: string,
-    deidServiceName: string,
-    options?: DeidServicesDeleteOptionalParams,
-  ) => PollerLike<OperationState<void>, void>;
+    options?: DeidServicesGetOptionalParams,
+  ) => Promise<DeidService>;
 }
 
-export function getDeidServices(context: HealthDataAIServicesContext, subscriptionId: string) {
+function _getDeidServices(context: HealthDataAIServicesContext) {
   return {
-    get: (
+    delete: (
       resourceGroupName: string,
       deidServiceName: string,
-      options?: DeidServicesGetOptionalParams,
-    ) => deidServicesGet(context, subscriptionId, resourceGroupName, deidServiceName, options),
-    listByResourceGroup: (
-      resourceGroupName: string,
-      options?: DeidServicesListByResourceGroupOptionalParams,
-    ) => deidServicesListByResourceGroup(context, subscriptionId, resourceGroupName, options),
-    listBySubscription: (options?: DeidServicesListBySubscriptionOptionalParams) =>
-      deidServicesListBySubscription(context, subscriptionId, options),
+      options?: DeidServicesDeleteOptionalParams,
+    ) => $delete(context, resourceGroupName, deidServiceName, options),
     create: (
       resourceGroupName: string,
       deidServiceName: string,
       resource: DeidService,
       options?: DeidServicesCreateOptionalParams,
-    ) =>
-      deidServicesCreate(
-        context,
-        subscriptionId,
-        resourceGroupName,
-        deidServiceName,
-        resource,
-        options,
-      ),
-    update: (
+    ) => create(context, resourceGroupName, deidServiceName, resource, options),
+    listBySubscription: (options?: DeidServicesListBySubscriptionOptionalParams) =>
+      listBySubscription(context, options),
+    listByResourceGroup: (
+      resourceGroupName: string,
+      options?: DeidServicesListByResourceGroupOptionalParams,
+    ) => listByResourceGroup(context, resourceGroupName, options),
+    get: (
       resourceGroupName: string,
       deidServiceName: string,
-      properties: DeidUpdate,
-      options?: DeidServicesUpdateOptionalParams,
-    ) =>
-      deidServicesUpdate(
-        context,
-        subscriptionId,
-        resourceGroupName,
-        deidServiceName,
-        properties,
-        options,
-      ),
-    delete: (
-      resourceGroupName: string,
-      deidServiceName: string,
-      options?: DeidServicesDeleteOptionalParams,
-    ) => deidServicesDelete(context, subscriptionId, resourceGroupName, deidServiceName, options),
+      options?: DeidServicesGetOptionalParams,
+    ) => get(context, resourceGroupName, deidServiceName, options),
   };
 }
 
-export function getDeidServicesOperations(
+export function _getDeidServicesOperations(
   context: HealthDataAIServicesContext,
-  subscriptionId: string,
 ): DeidServicesOperations {
   return {
-    ...getDeidServices(context, subscriptionId),
+    ..._getDeidServices(context),
   };
 }