+++ sdk/healthdataaiservices/arm-healthdataaiservices/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
-    
+
+## 1.1.0 (Unreleased)
+
+### Breaking Changes
+
+- Removed `beginUpdateAndWait` and `beginUpdate` methods from `DeidServicesOperations`
+- Removed model `DeidUpdate`
+- Removed model `DeidPropertiesUpdate`
+
 ## 1.0.0 (2024-11-25)
 
 ### Features Added